[Rhesus.Toolkit] Move utilities to Rhesus [Rhesus.Messenger] Add read/write methods for some missing data types [LRT14] Add profiler code [Dashboard 2.0] Add basic tab support; update later to match tab columns. [Dashboard 2.0] Start work on telemetry sending code (robot side)

+++ LRT14/Brain/Automation/Collect.cpp
@@ -44,9 +44,9 @@ bool Collect::Run()
 		bool wrongBall = false;
 		if (RobotState::Instance().Alliance() == DriverStation::kRed)
 		{
-			if (abs(m_redChannel->GetAverageValue() - m_blueBallRed) <= m_colorRange &&
-					abs(m_blueChannel->GetAverageValue() - m_blueBallBlue) <= m_colorRange &&
-					abs(m_greenChannel->GetAverageValue()- m_blueBallGreen) <= m_colorRange)
+			if (abs((int)m_redChannel->GetAverageValue() - m_blueBallRed) <= m_colorRange &&
+					abs((int)m_blueChannel->GetAverageValue() - m_blueBallBlue) <= m_colorRange &&
+					abs((int)m_greenChannel->GetAverageValue()- m_blueBallGreen) <= m_colorRange)
 			{
 				m_collectorRollers->SetDirection(CollectorRollersData::REVERSE);
 				wrongBall = true;
@@ -54,9 +54,9 @@ bool Collect::Run()
 		}
 		if (RobotState::Instance().Alliance() == DriverStation::kBlue)
 		{
-			if (abs(m_redChannel->GetAverageValue() - m_redBallRed) <= m_colorRange &&
-					abs(m_blueChannel->GetAverageValue() - m_redBallBlue) <= m_colorRange &&
-					abs(m_greenChannel->GetAverageValue()- m_redBallGreen) <= m_colorRange)
+			if (abs((int)m_redChannel->GetAverageValue() - m_redBallRed) <= m_colorRange &&
+					abs((int)m_blueChannel->GetAverageValue() - m_redBallBlue) <= m_colorRange &&
+					abs((int)m_greenChannel->GetAverageValue()- m_redBallGreen) <= m_colorRange)
 			{
 				m_collectorRollers->SetDirection(CollectorRollersData::REVERSE);
 				wrongBall = true;
